+++ lib/scss_lint/reporter/default_reporter.rb
@@ -1,13 +1,11 @@
-require 'colorize'
-
 module SCSSLint
   # Reports a single line per lint.
   class Reporter::DefaultReporter < Reporter
     def report_lints
       if lints.any?
         lints.map do |lint|
-          type = lint.error? ? '[E]'.red : '[W]'.yellow
-          "#{lint.filename.cyan}:" << "#{lint.line}".magenta <<
+          type = lint.error? ? '[E]'.color(:red) : '[W]'.color(:yellow)
+          "#{lint.filename.color(:cyan)}:" << "#{lint.line}".color(:magenta) <<
                                       " #{type} #{lint.description}"
         end.join("\n") + "\n"
       end